修改删除按钮手势冲突／隐藏按钮显示提示视图 修改emoji数据类型，导致数据错误 修改默认分组名称

+++ 表情键盘/Emoticon/NSString+HMEmoji.h
@@ -12,7 +12,7 @@
 /**
  *  将十六进制的编码转为emoji字符
  */
-+ (NSString *)emojiWithIntCode:(unsigned long)intCode;
++ (NSString *)emojiWithIntCode:(unsigned int)intCode;
 
 /**
  *  将十六进制的编码转为emoji字符
